+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## NEXT VERSION
+- Throwing an exception when command invoked by `EksctlCommand.run_and_get_output` or `AwsCommand.run_and_get_output` fails
+
 ## Version 1.0.8 - Feature and bugfix release
 - Add option to install Metrics Server
 - Fix "Inspect node pools" macro when using managed node groups